reduce dependency on header file
