+++ src/lightgbm_R.cpp
@@ -1,5 +1,6 @@
 #include <LightGBM/lightgbm_R.h>
 
+#include <LightGBM/utils/log.h>
 #include <LightGBM/utils/openmp_wrapper.h>
 #include <LightGBM/utils/text_reader.h>
 #include <LightGBM/utils/common.h>
